﻿Exercise 45:What Color is that Square?
Positions on a chess board are identified by a letter and a number. The letter identifies
the column, while the number identifies the row, as shown below:
1 a
2 b
3 c
4 d
5 e
6 f
7 g
8 h

Write a program that reads a position from the user. Use an if statement to determine if the column begins with a black square or a white square. Then use modular
arithmetic to report the color of the square in that row. For example, if the user enters
a1 then your program should report that the square is black. If the user enters d5
then your program should report that the square is white. Your program may assume
that a valid position will always be entered. It does not need to perform any error
checking.