Exercise 93: Next Prime
In this exercise you will create a function named next Prime that finds and returns
the first prime number larger than some integer, n. The value of n will be passed to the function as its only parameter. Include a main program that reads an integer from
the user and displays the first prime number larger than the entered value. Import
and use your solution to Exercise 92 while completing this exercise.